pyproject.toml
[tool.poetry]
name = "ExploreSCIS2021-Workshop"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.2"
jwt = "^1.3.1"
Flask-SSLify = "^0.1.5"
replit = "^3.2.4"
argon2-cffi = "^21.3.0"
Flask-API = "^3.0.post1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
